+++ src/ships/mod.rs
@@ -8,7 +8,7 @@ pub mod enemy;
 pub mod player;
 pub mod ship;
 
-use self::{bullet::BulletPlugin, player::PlayerPlugin};
+use self::player::PlayerPlugin;
 
 pub struct ShipsPlugin;
 impl Plugin for ShipsPlugin {
@@ -16,7 +16,6 @@ impl Plugin for ShipsPlugin {
         app.add_plugins((
             RapierPhysicsPlugin::<NoUserData>::pixels_per_meter(1.0),
             PlayerPlugin,
-            BulletPlugin,
         ));
     }
 }